+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Add Work Items to Freshservice change when available
+
 ## 1.0.0 - 2018-12-07
 ### Added
 - Base project, initial structure.
@@ -9,3 +13,5 @@ All notable changes to this project will be documented in this file.
 - API service health endpoint
 - Basic auth for `/vss` subroutes
 - Multi-stage Dockerfile
+
+[Unreleased]: https://github.com/payvision-development/scribe/compare/v1.0.0...HEAD